feat(79387) Refatora funções para OOP e altera estrutura de pasta para DRE
