+++ sme_ptrf_apps/core/services/notificacao_services/__init__.py
@@ -4,7 +4,6 @@ from .notificacao_comentario_pc_service import notificar_comentario_pc
 from .notificacao_proximidade_inicio_periodo_pc_service import notificar_proximidade_inicio_periodo_prestacao_conta
 from .notificacao_inicio_periodo_prestacao_de_contas import notificar_inicio_periodo_prestacao_de_contas
 from .notificacao_pendencia_envio_prestacao_de_contas import notificar_pendencia_envio_prestacao_de_contas
-from .notificacao_prazo_para_acerto_apos_vencimento import notificar_prazo_para_acerto_apos_vencimento
 from .notificacao_prestacao_de_contas_devolvida_para_acertos import notificar_prestacao_de_contas_devolvida_para_acertos
 from .notificacao_proximidade_fim_periodo_pc_service import notificar_proximidade_fim_periodo_prestacao_conta
 from .notificacao_atraso_entrega_ajustes_prestacao_de_contas import notificar_atraso_entrega_ajustes_prestacao_de_contas
