+++ sme_ptrf_apps/core/tasks.py
@@ -200,7 +200,7 @@ def gerar_notificacao_pendencia_envio_prestacao_de_contas_async():
     time_limet=600,
     soft_time_limit=300
 )
-def gerar_notificacao_prazo_para_acerto_apos_vencimento_async():
+def gerar_notificacao_prazo_para_acerto_consolidado_dre_apos_vencimento_async():
     logger.info(f'Iniciando a geração de notificação prazo acerto apos vencimento async')
 
     from sme_ptrf_apps.core.services.notificacao_services import notificar_prazo_para_acerto_apos_vencimento
