+++ sme_ptrf_apps/dre/services/notificacao_service/__init__.py
@@ -1,1 +1,2 @@
 from sme_ptrf_apps.dre.services.notificacao_service import class_notificacao_service
+from sme_ptrf_apps.dre.services.notificacao_service import notificacao_prazo_para_acerto_apos_vencimento
