+++ sme_ptrf_apps/dre/tests/tests_services/tests_notificacoes_services/test_notificacoes_prazo_para_acerto_apos_vencimento_service.py
@@ -2,8 +2,9 @@ import pytest
 
 from freezegun import freeze_time
 
-from ....core.models import Notificacao
-from ....core.services.notificacao_services import notificar_prazo_para_acerto_apos_vencimento
+from .....core.models import Notificacao
+from .....dre.services.notificacao_service.notificacao_prazo_para_acerto_apos_vencimento import NotificacaoConsolidadoPrazoAcertoVencimento
+
 
 pytestmark = pytest.mark.django_db
 
@@ -11,7 +12,9 @@ pytestmark = pytest.mark.django_db
 @freeze_time("2022-11-22")
 def test_deve_notificar_tecnico_dre(usuario_tecnico_notificavel, associacao_a, consolidado_dre_devolucao_apos_acertos):
     assert not Notificacao.objects.exists()
-    notificar_prazo_para_acerto_apos_vencimento(enviar_email=True)
+    NotificacaoConsolidadoPrazoAcertoVencimento(
+        enviar_email=True
+    ).notificar_prazo_para_acerto_apos_vencimento()
     notificacao = Notificacao.objects.first()
     assert Notificacao.objects.count() == 1
     assert notificacao.tipo == Notificacao.TIPO_NOTIFICACAO_ALERTA
@@ -25,6 +28,16 @@ def test_deve_notificar_tecnico_dre(usuario_tecnico_notificavel, associacao_a, c
 @freeze_time("2022-11-22")
 def test_nao_deve_notificar_pois_nao_tem_acerto_fora_prazo(periodo_2021_4_pc_2021_06_18_a_2021_06_30, usuario_notificavel, associacao_a, consolidado_dre_devolucao_apos_acertos_dentro_do_prazo):
     assert not Notificacao.objects.exists()
-    notificar_prazo_para_acerto_apos_vencimento(enviar_email=True)
+    NotificacaoConsolidadoPrazoAcertoVencimento(
+        enviar_email=True
+    ).notificar_prazo_para_acerto_apos_vencimento()
     assert Notificacao.objects.count() == 0
 
+
+@freeze_time("2022-11-22")
+def test_nao_deve_notificar_pois_nao_tem_usuario(periodo_2021_4_pc_2021_06_18_a_2021_06_30, usuario_notificavel, associacao_a, consolidado_dre_devolucao_apos_acertos_dentro_do_prazo):
+    assert not Notificacao.objects.exists()
+    NotificacaoConsolidadoPrazoAcertoVencimento(
+        enviar_email=True
+    ).notificar_prazo_para_acerto_apos_vencimento()
+    assert Notificacao.objects.count() == 0
